+++ src/main/java/kitchenpos/application/TableGroupService.java
@@ -4,9 +4,9 @@ import kitchenpos.dao.OrderDao;
 import kitchenpos.dao.OrderTableDao;
 import kitchenpos.dao.TableGroupDao;
 import kitchenpos.domain.OrderStatus;
-import kitchenpos.domain.OrderTable;
 import kitchenpos.domain.OrderTables;
 import kitchenpos.domain.TableGroup;
+import kitchenpos.ordertable.domain.OrderTable;
 import kitchenpos.request.TableGroupCreateRequest;
 import kitchenpos.request.TableGroupUnitDto;
 import org.springframework.stereotype.Service;
