+++ src/test/java/kitchenpos/application/OrderServiceTest.java
@@ -6,6 +6,7 @@ import kitchenpos.dao.OrderTableDao;
 import kitchenpos.dao.ProductDao;
 import kitchenpos.domain.*;
 import kitchenpos.menugroup.domain.MenuGroup;
+import kitchenpos.ordertable.domain.OrderTable;
 import kitchenpos.product.domain.Product;
 import kitchenpos.request.OrderCreateRequest;
 import kitchenpos.request.OrderLineItemDto;
