+++ src/test/java/kitchenpos/application/TableGroupServiceTest.java
@@ -6,6 +6,8 @@ import kitchenpos.dao.OrderTableDao;
 import kitchenpos.dao.ProductDao;
 import kitchenpos.domain.*;
 import kitchenpos.menugroup.domain.MenuGroup;
+import kitchenpos.ordertable.application.TableService;
+import kitchenpos.ordertable.domain.OrderTable;
 import kitchenpos.product.domain.Product;
 import kitchenpos.request.MenuCreateRequest;
 import kitchenpos.request.MenuProductDto;
