+++ src/test/java/kitchenpos/domain/OrderTablesTest.java
@@ -1,5 +1,6 @@
 package kitchenpos.domain;
 
+import kitchenpos.ordertable.domain.OrderTable;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
