+++ src/test/java/kitchenpos/fixture/TableGroupFixtures.java
@@ -1,6 +1,6 @@
 package kitchenpos.fixture;
 
-import kitchenpos.domain.OrderTable;
+import kitchenpos.ordertable.domain.OrderTable;
 import kitchenpos.request.TableGroupCreateRequest;
 import kitchenpos.request.TableGroupUnitDto;
